GITLAB-1537 Move TaskIdDTO in task package + visibility enhancements
This enforce the usage of TaskGenerator for returning tasks to the caller
